+++ benches/recovery_scale.rs
@@ -64,7 +64,6 @@ impl ScaleConfig {
             .with_io_engine(IoEngine::Posix)
             .with_io_mode(IoMode::Buffered)
             .with_io_workers(1)
-            .with_io_concurrency(1)
             .with_l1_capacity(self.memory_bytes)
             .with_memory_limit(self.memory_limit_bytes)
             .with_statistics(false)